backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langgraph>=0.2.0",
    "pydantic>=2.7.0",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.9",
    "pyyaml>=6.0.2",
    "uvicorn>=0.30.0",
]
